requirements.txt
deta
httpx~=0.22.0
python-dotenv~=0.16.0
python-telegram-bot~=20.0a0